# Implementation pack/05-openapi.yaml
openapi: 3.0.3
info:
  title: AI CV Tailoring and Cover Letter Platform — Backend API
  description: >
    Draft specification for developer implementation. Covers CV ingestion,
    job post ingestion, matching, tailored CV generation, guided cover
    letter generation, and export. All generated content must carry
    evidence references — see evidenceReferences fields throughout.
  version: 0.1.0-draft
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      security: []
      summary: Create user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '409':
          description: Email already registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Authenticate and issue session/token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      description: >
        Standard error envelope returned by every non-2xx response.
        `details` is free-form and endpoint-specific — populate it with
        whatever helps the caller correct the request (offending field,
        allowed values, etc.), omit it when there's nothing further to add.
      required: [status, code, message, timestamp, path]
      properties:
        status: { type: integer, description: "HTTP status code, duplicated in the body for logging convenience" }
        code: { type: string, description: "Stable machine-readable error code, e.g. VALIDATION_ERROR" }
        message: { type: string, description: "Human-readable explanation" }
        timestamp: { type: string, format: date-time }
        path: { type: string, description: "Request path that produced the error" }
        details:
          type: object
          additionalProperties: true
          nullable: true
      example:
        status: 400
        code: VALIDATION_ERROR
        message: "Invalid file type. Only PDF and DOCX are supported."
        timestamp: "2026-08-03T10:00:00Z"
        path: "/api/v1/cvs"
        details:
          field: file
          reason: unsupported_mime_type
          allowed: ["application/pdf", "application/vnd.openxmlformats-officedocument.wordprocessingml.document"]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        accountStatus: { type: string, enum: [active, suspended, deleted] }
        createdAt: { type: string, format: date-time }
...